Struct zoom::vector::space::Box

pub struct Box<V> {
    pub origin: V,
    pub offset: V,
}

A Box with a center at origin and one of the corners created by offset

The box is aligned so that the face normals point along each axis.

Fields

origin

offset

Methods

impl<V> Box<V>

fn new(origin: V, offset: V) -> Self

fn space<D>(&self) -> D where V: Vector<D>, D: Float